rename and refactor sample options
